Fix dtype conversion, consider models with PEFT wrapper

+++ ludwig/utils/llm_quantization_utils.py
@@ -1,3 +1,4 @@
+import torch
 from bitsandbytes.functional import dequantize_4bit
 from bitsandbytes.nn.modules import Linear4bit
 from torch import nn
@@ -29,12 +30,16 @@ class Linear4BitToLinear(nn.Module):
 
         # Dequantize the weight and bias from the Linear4bit layer and perform an in-place tensor replacement to update
         # the weights and bias in the new Linear layer. This is done to avoid creating a new tensor and copying the
-        # data, which is slow.
+        # data, which is slow. The to() call is needed to ensure the new tensors have the same dtype as the original
+        # dequantized tensors, which is fp16. Otherwise, the new tensors will be fp32 by default.
         new_linear_layer.weight.data.copy_(
             dequantize_4bit(linear4bit_layer.weight.data, linear4bit_layer.weight.quant_state)
         )
+        new_linear_layer.weight.data = new_linear_layer.weight.data.to(dtype=torch.float16)
+
         if linear4bit_layer.bias is not None:
             new_linear_layer.bias.data.copy_(linear4bit_layer.bias.data)
+            new_linear_layer.bias.data = new_linear_layer.bias.data.to(dtype=linear4bit_layer.bias.data.dtype)
 
         return new_linear_layer
